include <string.h> to pick up memcpy prototype. same. same. same.
2001-10-01  Charles Kerr <charles@rebelbase.com>

	* gmime-filter.c: include <string.h> to pick up memcpy prototype.
	* gmime-stream-mem.c: same.
	* gmime-stream-buffer.c: same.
	* gmime-stream-filter.c: same.

	* gmime-stream-mem.c (stream_flush): added a retval of 0.

	* gmime-filter-crlf.h (g_mime_filter_crlf_new_type): renamed
	prototype as g_mime_filter_crlf_new to sync with .c

	* gmime-part.h: added g_mime_part_set_content_object() prototype.

	* gmime-content-type.h: Replaced <> with "" in #include <gmime-param.h>

	* pan-mime-parser.c: added #include gmime-stream-buffer.h to
	pick up prototype for g_mime_stream_buffer_gets.
	* pan-mime-parser.c (g_strstrbound): removed unused func.

	* gmime-utils.c: on calls to ctype functions, explicitly upcast the
	char arguments as ints to shut up gcc warnings on Solaris.
	* gmime-param.c: same.

+++ gmime-utils.c
@@ -160,10 +160,10 @@ datetok (const gchar *date)
 	start = (gchar *) date;
 	while (*start) {
 		/* kill leading whitespace */
-		for ( ; *start && isspace (*start); start++);
+		for ( ; *start && isspace ((int)*start); start++);
 		
 		/* find the end of this token */
-		for (end = start; *end && !isspace (*end); end++);
+		for (end = start; *end && !isspace ((int)*end); end++);
 		
 		token = g_strndup (start, (end - start));
 		
@@ -233,7 +233,7 @@ get_mday (gchar *str)
 	g_return_val_if_fail (str != NULL, -1);
 	
 	for (p = str; *p; p++)
-		if (!isdigit (*p))
+		if (!isdigit ((int)*p))
 			return -1;
 	
 	mday = atoi (str);
@@ -265,7 +265,7 @@ get_year (const gchar *str)
 	const gchar * p;
 	
 	for (p = str; *p; p++)
-		if (!isdigit (*p))
+		if (!isdigit ((int)*p))
 			return -1;
 	
 	year = atoi (str);
@@ -289,7 +289,7 @@ get_time (const gchar *in, gint *hour, gint *min, gint *sec)
 	for (p = in; *p && digits; p++) {
 		if (*p == ':')
 			colons++;
-		else if (!isdigit (*p))
+		else if (!isdigit ((int)*p))
 			digits = FALSE;
 	}
